requirements.txt

torch
torchvision
transformers
segment-anything
numpy
opencv-python-headless
Pillow
google-generativeai
flask
pyngrok